Refactor: code splitting and media query on header
- Separate style codes from function components
- Make responsive header
