+++ components/layouts/Footer.tsx
@@ -1,16 +1,19 @@
-import styled from '@emotion/styled';
 import Image from 'next/image';
-import { CONTENT_MAX_WIDTH, MEDIA_QUERY_END_POINT } from '../../constants';
-import { CUSTOMER_SERVICE } from '../../constants/Footer';
+import {
+  COPYRIGHT,
+  CUSTOMER_SERVICE,
+  FOOTER_TERMS,
+} from '../../constants/Footer';
 import Logo from './../../public/ala_logo.png';
 import FooterCard from './FooterCard';
+import * as S from './Footer.style';
 
 const Footer = () => {
   return (
-    <Container>
-      <CSSection>
+    <S.Container>
+      <S.CSSection>
         <h2 className="a11y-hidden">고객지원 서비스</h2>
-        <UlElement>
+        <S.UlElement>
           {CUSTOMER_SERVICE.map(content => {
             return (
               <li key={content.ID}>
@@ -18,18 +21,18 @@ const Footer = () => {
               </li>
             );
           })}
-        </UlElement>
-      </CSSection>
-      <TermSection>
+        </S.UlElement>
+      </S.CSSection>
+      <S.TermSection>
         <h2 className="a11y-hidden">이용약관 및 회사SNS</h2>
-        <MaxWidthContainer className="container-wrapper">
+        <S.MaxWidthContainer className="container-wrapper">
           <ul className="container_terms">
-            <li>회사소개</li>
-            <li>이용약관</li>
-            <li>개인정보처리방침</li>
-            <li>멤버십 약관</li>
-            <li>공지사항</li>
-            <li>고객센터</li>
+            <li>{FOOTER_TERMS.ITEM_1}</li>
+            <li>{FOOTER_TERMS.ITEM_2}</li>
+            <li>{FOOTER_TERMS.ITEM_3}</li>
+            <li>{FOOTER_TERMS.ITEM_4}</li>
+            <li>{FOOTER_TERMS.ITEM_5}</li>
+            <li>{FOOTER_TERMS.ITEM_6}</li>
           </ul>
           <ul className="container_sns">
             <li>네이버블로그</li>
@@ -37,11 +40,11 @@ const Footer = () => {
             <li>인스타</li>
             <li>유튜브</li>
           </ul>
-        </MaxWidthContainer>
-      </TermSection>
-      <CompanyInfoSection>
+        </S.MaxWidthContainer>
+      </S.TermSection>
+      <S.CompanyInfoSection>
         <h2 className="a11y-hidden">사업자정보</h2>
-        <MaxWidthContainer className="wrapper">
+        <S.MaxWidthContainer className="wrapper">
           <dl>
             <div>
               <dt>카페드유라 대표이사 :</dt>
@@ -72,94 +75,14 @@ const Footer = () => {
               <dd>제2013-용인수지-00107호</dd>
             </div>
           </dl>
-          <CopyrightContainer>
+          <S.CopyrightContainer>
             <Image src={Logo} width={160} height={36} alt="" />
-            <p>&copy; CAFE de JuRA all rights reserved.</p>
-          </CopyrightContainer>
-        </MaxWidthContainer>
-      </CompanyInfoSection>
-    </Container>
+            <p>{COPYRIGHT}</p>
+          </S.CopyrightContainer>
+        </S.MaxWidthContainer>
+      </S.CompanyInfoSection>
+    </S.Container>
   );
 };
 
 export default Footer;
-
-const Container = styled.footer`
-  margin-top: 150px;
-  @media screen and (max-width: ${MEDIA_QUERY_END_POINT.MOBILE}) {
-    margin-top: 52px;
-  }
-`;
-
-const MaxWidthContainer = styled.div`
-  max-width: ${CONTENT_MAX_WIDTH};
-  margin: 0 auto;
-`;
-
-const UlElement = MaxWidthContainer.withComponent('ul');
-
-const CSSection = styled.section`
-  padding-bottom: 50px;
-  & > ul {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    margin: 0 auto;
-    & > li {
-      padding: 1rem;
-    }
-  }
-`;
-
-const TermSection = styled.section`
-  height: 80px;
-  background-color: #f5f5f5;
-  ul {
-    display: grid;
-    gap: 1rem;
-    li {
-      font-weight: bold;
-      font-size: 0.875rem;
-    }
-  }
-  div.container-wrapper {
-    display: flex;
-    justify-content: space-between;
-    margin: 0 auto;
-    height: 100%;
-    align-items: center;
-  }
-  ul.container_terms {
-    grid-template-columns: repeat(6, 1fr);
-  }
-  ul.container_sns {
-    grid-template-columns: repeat(4, 1fr);
-  }
-`;
-
-const CompanyInfoSection = styled.section`
-  height: 165px;
-  font-size: 0.875rem;
-  dl > div {
-    display: flex;
-  }
-  div.wrapper {
-    display: flex;
-    height: 100%;
-    align-items: center;
-    justify-content: space-between;
-  }
-  div.contact {
-    display: flex;
-    div {
-      display: flex;
-    }
-  }
-`;
-
-const CopyrightContainer = styled.div`
-  text-align: right;
-  p {
-    margin-top: 0.5rem;
-    font-size: 0.875rem;
-  }
-`;
