+++ components/layouts/Navigation.tsx
@@ -1,23 +1,26 @@
-import styled from '@emotion/styled';
 import Image from 'next/image';
 import Logo from './../../public/ala_logo.png';
-import { MEDIA_QUERY_END_POINT } from '../../constants';
 import Cart from './../../public/icon/cart.svg';
 import Account from './../../public/icon/account.svg';
+import Hamburger from './../../public/icon/hamburger-menu.svg';
 import { NAV_TAB_NAME } from '../../constants/Gnb';
+import Link from 'next/link';
+import * as S from './Navigation.style';
 
 const Navigation = () => {
   return (
-    <Header>
-      <ImgContainer>
-        <Image
-          src={Logo}
-          fill={true}
-          sizes="(max-width: 768px) 90vw"
-          alt="알라카르테"
-        />
-      </ImgContainer>
-      <Nav>
+    <S.Header>
+      <Link href="/">
+        <S.ImgContainer>
+          <Image
+            src={Logo}
+            fill={true}
+            sizes="(max-width: 768px) 90vw"
+            alt="알라카르테"
+          />
+        </S.ImgContainer>
+      </Link>
+      <S.Nav>
         <ul className="container_menu">
           <li>{NAV_TAB_NAME.TAB_1}</li>
           <li>{NAV_TAB_NAME.TAB_2}</li>
@@ -26,7 +29,7 @@ const Navigation = () => {
           <li>{NAV_TAB_NAME.TAB_5}</li>
         </ul>
         <ul className="container_icon-menu">
-          <li>
+          <li className="icon_cart">
             <Image src={Cart} alt={NAV_TAB_NAME.TAB_6} width={28} height={28} />
           </li>
           <li>
@@ -37,57 +40,13 @@ const Navigation = () => {
               height={25}
             />
           </li>
+          <li className="icon_menu">
+            <Image src={Hamburger} alt="메뉴" width={25} height={25} />
+          </li>
         </ul>
-      </Nav>
-    </Header>
+      </S.Nav>
+    </S.Header>
   );
 };
 
 export default Navigation;
-
-const Header = styled.header`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0 2.5rem;
-  height: 100px;
-  @media screen and (max-width: ${MEDIA_QUERY_END_POINT.MOBILE}) {
-    height: 52px;
-  }
-`;
-
-const ImgContainer = styled.div`
-  position: relative; // Image 컴포넌트의 fill 속성 사용하기 위해 설정
-  min-width: 124px;
-  height: 28px;
-`;
-
-const Nav = styled.nav`
-  width: 100%;
-  margin-left: 2rem;
-  display: flex;
-  justify-content: space-between;
-  ul {
-    display: grid;
-    li {
-      font-weight: bold;
-      font-size: 1.1rem;
-    }
-  }
-  ul.container_menu {
-    width: 90%;
-    grid-template-columns: repeat(5, 1fr);
-    text-align: center;
-  }
-  ul.container_icon-menu {
-    width: 10%;
-    min-width: 100px;
-    grid-template-columns: repeat(2, 1fr);
-    li {
-      text-align: right;
-    }
-  }
-  @media screen and (max-width: ${MEDIA_QUERY_END_POINT.MOBILE}) {
-    border: solid 1px violet;
-  }
-`;
